Game now finishes, added decent fragment navigation, cleaned up the fragments code

+++ project/app/src/main/java/edu/ub/pis2019/pis_16/tempac/Model/Score.kt
@@ -2,9 +2,9 @@ package edu.ub.pis2019.pis_16.tempac.Model
 
 class Score {
 
-    var score : Int = 0
+    var score : Float = 0f
 
-    fun update(bonus : Int){
+    fun update(bonus : Float){
         score += 1 + bonus
     }
 }
